Warnings clean-up, unnecessary casts.
git-svn-id: https://svn.apache.org/repos/asf/directory/apacheds/trunk@542339 13f79535-47bb-0310-9956-ffa450edef68

+++ protocol-dns/src/main/java/org/apache/directory/server/dns/io/encoder/DnsMessageEncoder.java
@@ -207,7 +207,7 @@ public class DnsMessageEncoder
     {
         if ( recursionDesired )
         {
-            return ( byte ) ( ( byte ) 0x01 );
+            return ( byte ) 0x01;
         }
         return 0;
     }
